I have moved all sorting logic to service, so that component will be more testable and focused on supporting the view.
